% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{isTrueVal}
\alias{isTrueVal}
\title{Check if an element contains a TRUE value}
\usage{
isTrueVal(variableName, variable, pos = as.integer(1))
}
\arguments{
\item{variableName}{The name of the variable that contains selected values.}

\item{variable}{The variable that contains selected values.}

\item{pos}{The list element to check.}
}
\value{
logical
}
\description{
This function checks the value returned from an mdqe
xpath selector.
}
\details{
If the variable is not defined or the value is \code{NA}, then \code{FALSE}
is returned. \code{TRUE} is only returned if the value is present and is \code{TRUE}.
}
