% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{ns_strip}
\alias{ns_strip}
\title{Strip the specified namespace from a document}
\usage{
ns_strip(x, nsPrefix)
}
\arguments{
\item{x}{The document to strip}

\item{nsPrefix}{The namespace prefix that identifies the}
}
\description{
Strip the specified namespace from a document
}
