% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{runCheck}
\alias{runCheck}
\title{Execute a quality check for a metadata document}
\usage{
runCheck(checkXML, metadataFile, sysmetaXML, checkFunction)
}
\arguments{
\item{checkXML}{(character) The filepath for a quality check document.}

\item{metadataFile}{(character) The filepath for a metadata document to check;
XML and JSON are currently supported.}

\item{sysmetaXML}{(character) The filepath for a system metadata document corresponding to the
metadata document, or a character string literal containing the system metadata.}

\item{checkFunction}{(not yet implemented)}
}
\value{
A named list of check results.
}
\description{
This function runs a quality check based on an XML document
that conforms to the format specified with the reference implementation
of the metadata quality engine: \url{https://github.com/NCEAS/metadig-engine}. The function
extracts values from the metadata document as specified in the quality check document
and runs the check R code to examine these values. The check returns a list providing information
about the success or failure of the check.
}
\examples{
checkXML <- system.file("extdata/dataset_title_length-check.xml", package = "metadig")
metadataFile <- system.file("extdata/example_EML.xml", package = "metadig")
sysmetaXML <- system.file("extdata/example_sysmeta.xml", package = "metadig")
result <- runCheck(checkXML, metadataFile, sysmetaXML)
}
