% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{selectNodes}
\alias{selectNodes}
\title{Select data values from a metadata document as specified in a checks selectors.}
\usage{
selectNodes(contextNode, selectorContext, selectorNamespaces)
}
\arguments{
\item{contextNode}{the metadata document xml2 node to apply the selector to.}

\item{selector}{the check document xml2 node of a selector to apply}

\item{selectorNamespace}{the XML namespaces that are defined for a selector}
}
\value{
an XML2 nodeset of the selected values (or XML2 nodes) from the metadata document
}
\description{
Select data values from a metadata document as specified in a checks selectors.
}

