% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_status}
\alias{set_status}
\title{helpers.R}
\usage{
set_status(status)
}
\arguments{
\item{status}{(character) The status message to set.}
}
\value{
This function is invoked for its side-effect.
}
\description{
General helpers for the check writing that are not explicitly checks.
Set the status in a check.
}

