% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{success}
\alias{success}
\title{Set the status to SUCCESS and, optionally, set an output message.}
\usage{
success(message = NULL)
}
\arguments{
\item{message}{(character) Optional. A message to set as an output.}
}
\value{
This function is invoked for its side-effect.
}
\description{
Set the status to SUCCESS and, optionally, set an output message.
}

