% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_presence}
\alias{check_presence}
\title{checks.R}
\usage{
check_presence(x)
}
\arguments{
\item{x}{(any) Metadata element to check.}
}
\value{
This function is invoked for its side-effect.
}
\description{
Utility functions to write checks with.
Check for the presence of an element. Specifically, this function returns
a SUCCESS status if one or more elements were present and a FAILURE status
if no elements were present.
}
\details{
Note: Modifies global environment!
}
