% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_presence}
\alias{check_presence}
\title{Check for the presence of an element}
\usage{
check_presence(x)
}
\arguments{
\item{x}{(any) Metadata element to check.}
}
\value{
This function is invoked for its side-effect.
}
\description{
This function checks for the presence of an element in a metadata file.
Specifically, this function returns a SUCCESS status if one or more elements
are present and a FAILURE status if no elements are present.
}
\details{
Note: Modifies the global environment.
}
