% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{failure}
\alias{failure}
\title{Set the status to FAILURE and, optionally, set an output message.}
\usage{
failure(message = NULL)
}
\arguments{
\item{message}{(character) Optional. A message to set as an output.}
}
\value{
This function is invoked for its side-effect.
}
\description{
Set the status to FAILURE and, optionally, set an output message.
}
