% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{runCheck}
\alias{runCheck}
\title{Execute a quality check for a metadata document.}
\usage{
runCheck(checkXML, metadataXML, checkFunction)
}
\arguments{
\item{checkFunction}{(not yet implemented)}

\item{check}{a character string that contains the filename of a check document}

\item{metadata}{a character string that contains the filename of a metdata document to check}
}
\value{
The check results, as a named list
}
\description{
The Metadata Improvement and Guidance (MetaDIG) project has defined a methodology
for running tests to evaluate the conformance of a metadata document to a set of quality metrics,
using an automated metadata quality engine that runs a suite of checks for a single or collection of metadata documents.
The reference implementation of this quality engine has been written in Java. The source
repository for this engine is \url{https://github.com/NCEAS/mdqengine}. A data repository that
uses this quality engine is located at \url{https://arcticdata.io}.

The \emph{metadig} R package
is a partial implementation of this quality engine in R, intended to aid in authoring metadata
checks in R, executing (aka "running") single quality checks and check suites locally in the R
environment. When complete and tested, these checks can then be including in quality suites
running the Java reference implementation of the quality engine.
The \code{runCheck} function runs a single quality check for a specified metadata document.
}
\details{
The quality check is an XML document that conforms to the format specified with
the reference implementation of the metadata quality engine: \url{https://github.com/NCEAS/mdqengine}.
\code{runCheck} extracts values from the metadata document as specified in the quality check document
and runs the check R code to examine these values. The check returns a list providing information
about the sucess or failure of the check.
}
\examples{
checkXML <- system.file("extdata/dataset_title_length-check.xml", package="metadig")
metadataXML<- system.file("extdata/example_EML.xml", package="metadig")
result <- runCheck(checkXML, metadataXML)
}
\seealso{
\code{\url{https://github.com/NCEAS/mdqengine}}
}
\keyword{check}
\keyword{quality}
