% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSuite.R
\name{runSuite}
\alias{runSuite}
\title{Execute a quality suite for a metadata document}
\usage{
runSuite(suiteXML, dirXML, metadataXML, sysmetaXML)
}
\arguments{
\item{suiteXML}{(character) A filepath for the quality suite document.}

\item{dirXML}{(character) A filepath for the directory storing quality check documents.}

\item{metadataXML}{(character) A filepath for the metadata document to check.}

\item{sysmetaXML}{(character) A file path for the system metadata document corresponding to the metadata document}
}
\value{
A named list of results.
}
\description{
This function runs a suite of quality checks based on an XML document
that conforms to the format specified with the reference implementation
of the metadata quality engine: \url{https://github.com/NCEAS/mdqengine}.
}
\examples{
\dontrun{
metadataXML <- system.file("extdata/example_EML.xml", package = "metadig")
suiteXML <- system.file("extdata/example_suite.xml", package = "metadig")
dirXML <- system.file("extdata", package = "metadig")
sysmetaXML <- system.file("extdata/example_sysmeta.xml", package = "metadig")

results <- runSuite(suiteXML, dirXML, metadataXML, sysmetaXML)
}
}
