% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_output}
\alias{save_output}
\title{Save an object as an output}
\usage{
save_output(x, type = "text")
}
\arguments{
\item{x}{(any) Object to save as output.}

\item{type}{(character) The MIME type to save the object as.}
}
\value{
This function is invoked for its side-effect.
}
\description{
Save an object as an output.
}
\details{
Note: Modifies the global environment.
}
