% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_status}
\alias{set_status}
\title{Set the status in a check}
\usage{
set_status(status)
}
\arguments{
\item{status}{(character) The status message to set.}
}
\value{
This function is invoked for its side-effect.
}
\description{
This function sets the status in a check.
}
\details{
Repeated calls to this function within a single check has special behavior:
\itemize{
\item If the status is ever set to FAILURE, it cannot be set back to SUCCESS
\item If the status is ever set to SUCCESS, it can always be set to FAILURE
}
}
